+++ crates/ruma-client-api/CHANGELOG.md
@@ -83,6 +83,8 @@ Improvements:
 - Add support for the Policy Server public key information endpoint, according
   to MSC4284 / Matrix 1.18.
 - Add `M_USER_LIMIT_EXCEEDED` error code according to MSC4335.
+- Add support for the user suspension and locking endpoints, according to
+  MSC4323 / Matrix 1.18. 
 
 # 0.22.1